fix(map): 언마운트시 클린업 함수 수정

+++ apps/desserbee-web/src/app/[lang]/(user)/(menu)/(search)/map/_components/KakaoMap/index.tsx
@@ -491,12 +491,16 @@ export function KakaoMap({
 
   useEffect(() => {
     return () => {
-      if (servicesRef.current.geoService && servicesRef.current.mapService) {
+      if (
+        servicesRef.current.geoService &&
+        servicesRef.current.mapService &&
+        isInitialized
+      ) {
         servicesRef.current.geoService.stopWatchingPosition();
         servicesRef.current.mapService.removeCurrentPositionMarker();
       }
     };
-  }, [servicesRef]);
+  }, [servicesRef, isInitialized]);
 
   const handleRefetchBtnClick = () => {
     setIsFetchRequired(true);
